Add tabs
Add navigation on bottom of screen using the tabs layout in expo router: https://docs.expo.dev/router/advanced/tabs/
